move reuse s3 object to aws request
